Added DatabaseProviderConfig to configure which database provider the application should connect to. This should then fix Azure database issue, by using SqLite on deployed application.
