+++ .gitignore
@@ -475,3 +475,9 @@ $RECYCLE.BIN/
 
 # Windows shortcuts
 *.lnk
+
+# Ignore SQLITE Database files
+*.db
+*.db-shm
+*.db-wal
+
