+++ src/Chirp.Migrations/appsettings.Development.json
@@ -1,6 +1,6 @@
 {
   "ConnectionStrings": {
-    "ChirpDatabase": "Server=localhost, 1433;Database=ChirpDockerDB;User Id=SA;Password=adminAdmin!;Trust Server Certificate=true",
-    "ChirpDatabase-sqLite": "Data Source=/users/marcus/Desktop/Chirp/chirp-database.db"
+    "ChirpDatabase-SqlServer": "Server=localhost, 1433;Database=ChirpDockerDB;User Id=SA;Password=adminAdmin!;Trust Server Certificate=true",
+    "ChirpDatabase-SqLite": "Data Source=chirp-database.db"
   }
   }
