+++ src/Chirp.Migrations/appsettings.Production.json
@@ -1,6 +1,6 @@
 {
-    "ConnectionStrings": 
-    {
-      "ChirpDatabase": "Server=tcp:bdsagroup22-chirpserver.database.windows.net,1433;Initial Catalog=bdsagroup22-chirpdb;Persist Security Info=False;User ID=adminUser;Password=adminAdmin!;MultipleActiveResultSets=False;Encrypt=True;TrustServerCertificate=False;Connection Timeout=30;"
-    }
+  "ConnectionStrings": {
+    "ChirpDatabase-SqlServer": "Server=tcp:bdsagroup22-chirpserver.database.windows.net,1433;Initial Catalog=bdsagroup22-chirpdb;Persist Security Info=False;User ID=adminUser;Password=adminAdmin!;MultipleActiveResultSets=False;Encrypt=True;TrustServerCertificate=False;Connection Timeout=30;",
+    "ChirpDatabase-SqLite": "Data Source=chirp-database.db"
+  }
 }
