+++ src/Chirp.Migrations/appsettings.json
@@ -1,2 +1,5 @@
 {
+  "DatabaseProviderConfig": {
+    "DatabaseProviderType": "SqlServer"
+  }
 }
