+++ src/Chirp.Web/DatabaseProviderConfig.cs
@@ -1,0 +1,14 @@
+﻿using System;
+namespace Chirp.Web
+{
+	public enum DatabaseProviderType
+	{
+		SqLite, SqlServer
+	}
+
+	public class DatabaseProviderConfig
+	{
+		public DatabaseProviderType DatabaseProviderType { get; set; }
+	}
+}
+
