+++ src/Chirp.Web/appsettings.json
@@ -21,5 +21,8 @@
       "Microsoft.AspNetCore": "Warning"
     }
   },
-  "AllowedHosts": "*"
+  "AllowedHosts": "*",
+  "DatabaseProviderConfig": {
+     "DatabaseProviderType":  "SqlServer"
+  }
 }
